优化copierx 支持uint64到sql.nullint64互转

+++ copierx/copierx.go
@@ -98,7 +98,7 @@ var (
 				return nil, errors.New("src type not matching")
 			}
 			if !s.Valid {
-				return 0, nil
+				return uint64(0), nil
 			}
 			return uint64(s.Int64), nil
 		},
